layoutの継承の設定
