+++ front/chat/app/(auth)/login/page.tsx
@@ -1,0 +1,5 @@
+export default function Home() {
+  return (
+    <div>login</div>
+  )
+}
