+++ front/chat/app/(auth)/signup/page.tsx
@@ -1,0 +1,5 @@
+export default function Home() {
+  return (
+    <div>signup</div>
+  )
+}
