+++ front/chat/app/layout.tsx
@@ -2,8 +2,6 @@ import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
 import Header from "@/component/layout/header";
-import Sidebar from "@/component/layout/sidebar";
-import Box from "@mui/material/Box";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -21,17 +19,7 @@ export default function RootLayout({
     <html lang="ja">
       <body className={inter.className}>
         <Header />
-        <Box
-          sx={{
-            display: "flex",
-            flexWrap: "wrap",
-            bgcolor: "background.paper",
-            borderRadius: 1,
-          }}
-        >
-          <Sidebar />
-          {children}
-        </Box>
+        {children}
       </body>
     </html>
   );
